TASK: Refactor Html::BLOCK_TAGS
Renames Html::BLOCK_TAGS to Html::SPACES_TAGS being an array of HTML
elements that natively is warpped into spaces or is displayed as a
block-level element by rendering engines.
Html::BLOCK_TAGS suggested that the array contains HTML block tags,
being elements that consist of an opening and a closing tag
(in contast to void tags).

Moreover the array is made one-dimensional instead of a milti-dimensional.

Also removes commented-out parts.