BUGFIX: Fix patterns using the Html::ATTRIBUTES_PATTERN constant
Assures that the patterns expect either a `>` or the end of the string
after the attributes list, instead of usng `>?`.
Otherwise a string of `<thhead>` would match a pattern like
`<th{attr_list}>?`.

This also makes Html::parsePlaintext() to match plaintext tags without a
closing `>`, matching the whole part to the string's end.